Resolve "Ошибка при вычислении MILLSEC_INTERVAL"

Closes #362

See merge request BI/magreport2.1!357

+++ magreport-backend/src/main/java/ru/magnit/magreportbackend/metrics_function/impl/AvgIntegerFunction.java
@@ -4,14 +4,14 @@ import ru.magnit.magreportbackend.metrics_function.MetricsFunction;
 
 public class AvgIntegerFunction implements MetricsFunction {
 
-    private int numValues;
-    private int sumOfValues;
+    private long numValues;
+    private long sumOfValues;
 
     @Override
     public void addValue(String value, int col, int row) {
         if (value != null) {
             numValues++;
-            sumOfValues += Double.parseDouble(value);
+            sumOfValues += Long.parseLong(value);
         }
     }
